+++ src/Milax/Mconsole/Models/Page.php
@@ -9,4 +9,15 @@ class Page extends Model
 	
 	protected $fillable = ['slug', 'title', 'heading', 'preview', 'text', 'description', 'hide_heading', 'fullwidth', 'system', 'enabled'];
 	
+	/**
+	 * Transform updated_at proerty.
+	 * 
+	 * @access public
+	 * @return string
+	 */
+	public function getUpdatedAttribute()
+	{
+		return $this->updated_at->format('m.d.Y');
+	}
+	
 }