+++ src/Milax/Mconsole/Providers/MconsoleServiceProvider.php
@@ -22,6 +22,7 @@ class MconsoleServiceProvider extends ServiceProvider
 	{
 		require __DIR__ . '/../Http/routes.php';
 		
+		// Resources
 		$this->loadTranslationsFrom(__DIR__ . '/../../../resources/lang', 'mconsole');
 		$this->loadViewsFrom(__DIR__ . '/../../../resources/views', 'mconsole');
 		
@@ -38,15 +39,6 @@ class MconsoleServiceProvider extends ServiceProvider
 				$migrations[$dir . $file] = base_path('database/migrations/' . $file);
 		});
 		$this->publishes($migrations, 'migrations');
-		
-		// Database seeds
-		$seeds = [];
-		$dir = __DIR__ . '/../../../seeds/';
-		collect(scandir(__DIR__ . '/../../../seeds'))->each(function ($file) use (&$dir, &$seeds) {
-			if (strpos($file, '.php') !== false)
-				$seeds[$dir . $file] = base_path('database/seeds/' . $file);
-		});
-		$this->publishes($seeds, 'seeds');
 	}
 
 	/**