+++ src/resources/lang/en/sections.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'users' => 'Users',
+    'list' => 'List',
+    'create' => 'Create',
+];