+++ src/resources/lang/ru/mconsole.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+	'text' => [
+		'welcome' => 'Добро пожаловать в Mconsole!',
+		'version' => 'версия :version',
+	],
+	'status' => [
+		'created' => 'Запись создана.',
+		'updated' => 'Запись обновлена.',
+		'deleted' => 'Запись удалена.',
+	],
+];