+++ CHANGELOG.md
@@ -1,7 +1,24 @@
+## 0.0.9
+
+Added:
+  - Added Users section
+  - Added first Blade directive
+  - Added localization
+  - Added system template
+  - Added `redirect` method for cms controllers (must have `protected $redirectTo` property)
+  - Added version display in dashboard
+  - Added admin creation prompt when installing mconsole
+
+Fixed:
+  - Fixed installer error while seeding database
+
 ## 0.0.8
 
-  - Package self installer `php artisan mconsole:install`
+Added:
+  - Added package self installer `php artisan mconsole:install`
+  - Added `BladeRenderer` class
+  - Added models: MconsoleRole, MconsoleMenu, MconsoleUser
+
+Updated:
   - Models not publishing to app anymore
   - Views moved in `views/mconsole` directory
-  - `BladeRenderer` class
-  - Added models: MconsoleRole, MconsoleMenu, MconsoleUser